added basic bot
pulled a lot of functionality into the Player class and I treat the bot just like the player in terms of position updates, I only set what inputs it is pressing. Master sends info on bot position to other peers. So far no collision detection on the bot and only a basic "following" state